requirements.txt
Flask
pdfplumber
pytesseract
fuzzywuzzy
Pillow
